Add independent platform support
