+++ src/app/layout.tsx
@@ -5,7 +5,10 @@ import './globals.css'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'jpegs',
+  title: {
+    template: "%s | jpegs",
+    default: "jpegs",
+  },
   description: 'Download photos and videos',
 }
