+++ src/components/footer.tsx
@@ -1,0 +1,10 @@
+import { FC } from "react";
+
+const Footer: FC = () => (
+  <footer className="mt-8 mb-8 container mx-auto text-center text-slate-400 text-sm">
+    <hr />
+    <div className="mt-2">@ 2024 jpegs. All rights reserved.</div>
+  </footer>
+);
+
+export default Footer;
