+++ src/utils.ts
@@ -83,3 +83,5 @@ export const findHighestQualityTwitterVideo = (item: TwitterVideo) => {
     }
     return best;
 };
+
+export const capitalize = (s: string) => s && s[0].toUpperCase() + s.slice(1);
